UX: remap native emojis in composer preview window properly.
Currently, the native emojis like `😀` are remapped only after the server side post cooking. After this commit it will get remapped in the composer preview window itself.
