+++ lib/pretty_text/shims.js
@@ -3,6 +3,7 @@ __buildOptions = require("pretty-text/pretty-text").buildOptions;
 __performEmojiUnescape = require("pretty-text/emoji").performEmojiUnescape;
 __emojiReplacementRegex = require("pretty-text/emoji").emojiReplacementRegex;
 __performEmojiEscape = require("pretty-text/emoji").performEmojiEscape;
+__emojiUnicodeReplacer = require("pretty-text/emoji").emojiUnicodeReplacer;
 __resetTranslationTree = require("pretty-text/engines/discourse-markdown/emoji")
   .resetTranslationTree;
 
@@ -24,47 +25,6 @@ define("discourse-common/lib/helpers", ["exports"], function (exports) {
 });
 
 __utils = require("discourse/lib/utilities");
-
-__emojiUnicodeReplacer = null;
-
-__setUnicode = function (replacements) {
-  const regexp = new RegExp(__emojiReplacementRegex, "g");
-
-  __emojiUnicodeReplacer = function (text) {
-    regexp.lastIndex = 0;
-
-    let m;
-    while ((m = regexp.exec(text)) !== null) {
-      let match = m[0];
-
-      let replacement = replacements[match];
-
-      if (!replacement) {
-        // if we can't find replacement for an emoji match
-        // attempts to look for the same without trailing variation selector
-        match = match.replace(/\ufe0f$/g, "");
-        replacement = replacements[match];
-      }
-
-      if (!replacement) {
-        continue;
-      }
-
-      replacement = ":" + replacement + ":";
-      const before = text.charAt(m.index - 1);
-      if (!/\B/.test(before)) {
-        replacement = "\u200b" + replacement;
-      }
-      text = text.replace(match, replacement);
-    }
-
-    // fixes Safari VARIATION SELECTOR-16 issue with some emojis
-    // https://meta.discourse.org/t/emojis-selected-on-ios-displaying-additional-rectangles/86132
-    text = text.replace(/\ufe0f/g, "");
-
-    return text;
-  };
-};
 
 __paths = {};
